send/recv ui
